wlan: Suspend link for sta run time channel switch
When channel switch timer expires, driver checks if scan
or ROC is in progress and ignores the channel switch request
if ROC or scan in progress. CB change is checked during beacon
processing and channel switch request is sent to firmware
without suspending the link. ROC request is processed after
this which suspends the link by sending init_scan request to
firmware. On receiving channel switch rsp for CB change, finish
scan req is sent as part of resume link since link is suspended.
init_scan fails resulting in setting lim scan state to idle.
finish_scan rsp is not processed properly since lim scan state
is idle instead of resume_link wait state and resume callback is
not set to NULL. As a result, resume link is not processed for
subsequent 2 ROC and finish_scan request is not sent to firmware.
The third ROC req will cause random mac addr count to exceed the
maximum allowed in firmware.

Fix is to suspend link as part of sta run time channel switch.

Change-Id: I9bf553c0bf9639ee581913ce2f210459370fddcc
CRs-Fixed: 2251186